requirements.txt
Flask-SQLAlchemy
Flask-Bcrypt
Flask-JWT-Extended
python-dotenv
Flask-Mail
Flask-WTF